[ALPS04655409] boost_ctrl: fix coverity
[Detail] fix coverity

MTK-Commit-Id: 273cc3210a565eb2ff5269e73873d4a5eed3ff2b

CR-Id: ALPS04655409
Feature: [Module]PowerHAL

+++ drivers/misc/mediatek/performance/fpsgo_v3/fbt/src/mini_top.c
@@ -371,7 +371,7 @@ static int __minitop_has_heavy(struct minitop_rec *mr, int *heavy)
 			keep |= 1;
 		else
 			minitop_trace(" %5d %3d%% end debnc FBT",
-				      mr->tid, mr->ratio_inst);
+				      mr->tid, (int)mr->ratio_inst);
 	}
 
 	if ((mr->source & MINITOP_SCHED) == 0)
@@ -402,7 +402,7 @@ end_sched_nomi:
 	ret = keep ? 0 : -EAGAIN;
 	if (mr->ever)
 		minitop_trace(" %5d %3d%% end debnc heavy",
-			      mr->tid, mr->ratio);
+			      mr->tid, (int)mr->ratio);
 
 err_cleanup:
 	*heavy = 0;
